+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "cirrus-lib/crates/aiff-rs"]
-	path = cirrus-lib/crates/aiff-rs
-	url = https://github.com/lunarbridge/aiff-rs
+[submodule "crates/aiff-rs"]
+	path = crates/aiff-rs
+	url = https://github.com/fibremint/aiff-rs
