+++ cirrus-app/src-tauri/Cargo.toml
@@ -15,7 +15,7 @@ rust-version = "1.57"
 tauri-build = { version = "1.0.0-rc.4", features = [] }
 
 [dependencies]
-cirrus-tauri-plugin = { path = "../../cirrus-lib/crates/tauri-plugin" }
+cirrus-tauri-plugin = { path = "../../crates/cirrus-tauri-plugin" }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 tauri = { version = "1.0.0-rc.4", features = [] }
