--- cirrus-lib/crates/client/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "cirrus-client-lib"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1"
-#bytes = "1"
-#cirrus-grpc = { path = "../grpc", features = ["only-client"] }
-cirrus-grpc = { path = "../grpc" }
-cpal = "0.13"
-futures = "0.3"
-tokio = { version = "1.0", features = ["rt-multi-thread", "time", "macros", "fs"] }
-tonic = "0.6"
-serde = "1"
-#prost = "0.9"
-rubato = "0.12"
-ndarray = "0.15"
-
-#[build-dependencies]
-#tonic-build = "0.6"
