--- cirrus-lib/crates/grpc/.gitignore
@@ -1,1 +1,0 @@
-proto-rust/
