--- cirrus-lib/crates/grpc/src/lib.rs
@@ -1,9 +1,0 @@
-include!("../proto-rust/cirrus.rs");
-
-pub mod api {
-    include!("../proto-rust/cirrus.api.rs");
-}
-
-pub mod common {
-    include!("../proto-rust/cirrus.common.rs");
-}
