--- cirrus-lib/crates/server/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "cirrus-server-lib"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-#aiff = {git = "https://github.com/julientregoat/aiff-rs.git"}
-aiff = { path = "../aiff-rs" }
-bson = { version = "2.1", features = ["chrono-0_4"] }
-bytes = "1.1"
-chrono = { version = "0.4", features = ["serde"] }
-futures = "0.3"
-cirrus-grpc = { path = "../grpc" }
-http = "0.2"
-notify = "4"
-mongodb = { version = "2.1", default-features = false, features = ["tokio-runtime"] }
-serde = "1"
-# symphonia = "0.5"
-tonic = "0.6"
-tokio = { version = "1.0", features = ["rt-multi-thread", "time", "macros", "fs"] }
-tokio-stream = "0.1"
-prost = "0.9"
-walkdir = "2"
-
-[build-dependencies]
-tonic-build = "0.6"
