+++ cirrus-server/src/main.rs
@@ -1,8 +1,94 @@
+mod logic;
+mod model;
+mod service;
+mod util;
+
+use std::sync::Arc;
+use std::sync::mpsc::channel;
+use std::time::Duration;
+
+use futures::Future;
+use notify::{Watcher, RecursiveMode, watcher};
+use tokio::{sync::Mutex, task::JoinHandle};
+use tonic::transport::Server as TonicServer;
+
+use cirrus_protobuf::{
+    audio_data_svc_server::AudioDataSvcServer,
+    audio_library_svc_server::AudioLibrarySvcServer,
+    audio_tag_svc_server::AudioTagSvcServer,
+};
+
+use crate::model::get_mongodb_client;
+
+async fn grpc_server_task(addr: &str, mongodb_client: mongodb::Client) -> Result<(), Box<dyn std::error::Error>> {
+    println!("run grpc server");
+    let addr = addr.parse().unwrap();
+
+    // let audio_data_svc = service::AudioDataSvcImpl::default();
+    let audio_data_svc = service::AudioDataSvcImpl::new(mongodb_client.clone());
+    let audio_library_svc = service::AudioLibrarySvcImpl::new(mongodb_client.clone());
+    let audio_tag_svc = service::AudioTagSvcImpl::new(mongodb_client.clone());
+
+    TonicServer::builder()
+        .add_service(AudioDataSvcServer::new(audio_data_svc))
+        .add_service(AudioLibrarySvcServer::new(audio_library_svc))
+        .add_service(AudioTagSvcServer::new(audio_tag_svc))
+        .serve(addr)
+        .await?;
+
+    Ok(())
+}
+
+fn run_fs_notify() -> Result<(), ()> {
+    let (tx, rx) = channel();
+
+    let mut watcher = watcher(tx, Duration::from_secs(5)).unwrap();
+    watcher.watch("d:\\tmp", RecursiveMode::Recursive).unwrap();
+
+    loop {
+        match rx.recv() {
+            Ok(event) => println!("{:?}", event),
+            Err(e) => println!("watch error: {:?}", e),
+        }
+    }
+}
+
+// fn process_chaged_audio_files_at_start() {
+//     let audio_types = ["aiff"];
+
+//     // let mut libraries: HashMap<PathBuf, Vec<document::AudioFile>> = HashMap::new();
+//     // let mut libraries = HashSet::new();
+//     // let mut audio_file_docs: Vec<document::AudioFile> = Vec::new();
+
+//     let audio_file_entry_iter = WalkDir::new(library_root).into_iter()
+//         .filter_map(|item| item.ok())
+//         // .map(|item| item.unwrap())
+//         .filter(|item| 
+//             item.metadata().unwrap().is_file() && 
+//             item.path().extension() != None)
+//         .filter(|item| {
+//             let file_extension = item.path().extension().unwrap();
+//             audio_types.contains(&file_extension.to_str().unwrap())
+//         });
+// }
+
+pub async fn run_server(addr: &str) -> Result<(), Box<dyn std::error::Error>> {
+    // std::thread::spawn(|| {
+    //     run_fs_notify()
+    // });
+
+    let mongodb_client = get_mongodb_client().await?;
+
+    grpc_server_task(addr, mongodb_client.clone()).await?;
+
+    Ok(())
+}
+
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let addr = "127.0.0.1:50000";
 
-    cirrus_server_lib::run_server(addr).await?;
+    run_server(addr).await?;
 
     Ok(())
 }
