+++ cirrus-server/src/util/hash.rs
