+++ crates/cirrus-client-core/src/dto.rs
