+++ crates/cirrus-tauri-plugin/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cirrus-tauri-plugin"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+cirrus-client-core = { path = "../cirrus-client-core" }
+cirrus-protobuf = { path = "../cirrus-protobuf" }
+http = "0.2"
+tauri = { version = "1.0.0-rc.4", default-features = false, features = [] }
+tokio = { version = "1.0", features = ["rt-multi-thread", "time", "macros", "fs"] }
+serde = "1"
